Rename addon to influxdb overwritting previous one
