+++ bundles/org.openhab.persistence.influxdb/pom.xml
@@ -1,47 +1,136 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
-  <modelVersion>4.0.0</modelVersion>
+	<modelVersion>4.0.0</modelVersion>
 
-  <parent>
-    <groupId>org.openhab.addons.bundles</groupId>
-    <artifactId>org.openhab.addons.reactor.bundles</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-  </parent>
+	<parent>
+		<groupId>org.openhab.addons.bundles</groupId>
+		<artifactId>org.openhab.addons.reactor.bundles</artifactId>
+		<version>3.0.0-SNAPSHOT</version>
+	</parent>
 
-  <artifactId>org.openhab.persistence.influxdb</artifactId>
+	<artifactId>org.openhab.persistence.influxdb</artifactId>
 
-  <name>openHAB Add-ons :: Bundles :: Persistence Service :: InfluxDB</name>
+	<name>openHAB Add-ons :: Bundles :: Persistence Service :: InfluxDB</name>
 
-  <properties>
-    <bnd.importpackage>!android.*,!com.android.*,!com.google.appengine.*,!org.apache.harmony.*,!org.apache.http.*,!rx.*</bnd.importpackage>
-  </properties>
+	<properties>
+		<bnd.importpackage>
+			!javax.annotation;!android.*,!com.android.*,!com.google.appengine.*,!dalvik.system,!kotlin.*,!org.conscrypt,!sun.security.ssl,!org.apache.harmony.*,!org.apache.http.*,!rx.*
+		</bnd.importpackage>
+	</properties>
 
-  <dependencies>
-    <!-- https://mvnrepository.com/artifact/org.influxdb/influxdb-java -->
-    <dependency>
-      <groupId>org.influxdb</groupId>
-      <artifactId>influxdb-java</artifactId>
-      <version>2.2</version>
-    </dependency>
-    <!-- https://mvnrepository.com/artifact/com.squareup.okhttp/okhttp -->
-    <dependency>
-      <groupId>com.squareup.okhttp</groupId>
-      <artifactId>okhttp</artifactId>
-      <version>2.4.0</version>
-    </dependency>
-    <!-- https://mvnrepository.com/artifact/com.squareup.okio/okio -->
-    <dependency>
-      <groupId>com.squareup.okio</groupId>
-      <artifactId>okio</artifactId>
-      <version>1.4.0</version>
-    </dependency>
-    <!-- https://mvnrepository.com/artifact/com.squareup.retrofit/retrofit -->
-    <dependency>
-      <groupId>com.squareup.retrofit</groupId>
-      <artifactId>retrofit</artifactId>
-      <version>1.9.0</version>
-    </dependency>
-  </dependencies>
+	<dependencies>
+		<!-- START InfluxDB 2.0-->
+		<!-- START influxdb-client-java-->
+		<dependency>
+			<groupId>com.influxdb</groupId>
+			<artifactId>influxdb-client-java</artifactId>
+			<version>1.6.0</version>
+		</dependency>
+		<dependency>
+			<artifactId>influxdb-client-core</artifactId>
+			<groupId>com.influxdb</groupId>
+			<version>1.6.0</version>
+		</dependency>
+		<dependency>
+			<artifactId>converter-gson</artifactId>
+			<groupId>com.squareup.retrofit2</groupId>
+			<version>2.5.0</version>
+		</dependency>
+		<dependency>
+			<artifactId>converter-scalars</artifactId>
+			<groupId>com.squareup.retrofit2</groupId>
+			<version>2.5.0</version>
+		</dependency>
+		<dependency>
+			<artifactId>gson</artifactId>
+			<groupId>com.google.code.gson</groupId>
+			<version>2.8.5</version>
+		</dependency>
+		<dependency>
+			<artifactId>gson-fire</artifactId>
+			<groupId>io.gsonfire</groupId>
+			<version>1.8.0</version>
+		</dependency>
+		<dependency>
+			<artifactId>okio</artifactId>
+			<groupId>com.squareup.okio</groupId>
+			<version>1.17.3</version>
+		</dependency>
+		<dependency>
+			<artifactId>commons-csv</artifactId>
+			<groupId>org.apache.commons</groupId>
+			<version>1.6</version>
+		</dependency>
+		<dependency>
+			<artifactId>json</artifactId>
+			<groupId>org.json</groupId>
+			<version>20180813</version>
+		</dependency>
+		<dependency>
+			<artifactId>okhttp</artifactId>
+			<groupId>com.squareup.okhttp3</groupId>
+			<version>3.13.1</version>
+		</dependency>
+		<dependency>
+			<artifactId>retrofit</artifactId>
+			<groupId>com.squareup.retrofit2</groupId>
+			<version>2.5.0</version>
+		</dependency>
+		<dependency>
+			<artifactId>jsr305</artifactId>
+			<groupId>com.google.code.findbugs</groupId>
+			<version>3.0.2</version>
+		</dependency>
+		<dependency>
+			<artifactId>logging-interceptor</artifactId>
+			<groupId>com.squareup.okhttp3</groupId>
+			<version>3.13.1</version>
+		</dependency>
+		<dependency>
+			<artifactId>rxjava</artifactId>
+			<groupId>io.reactivex.rxjava2</groupId>
+			<version>2.2.17</version>
+		</dependency>
+		<dependency>
+			<artifactId>reactive-streams</artifactId>
+			<groupId>org.reactivestreams</groupId>
+			<version>1.0.3</version>
+		</dependency>
+		<dependency>
+			<artifactId>swagger-annotations</artifactId>
+			<groupId>io.swagger</groupId>
+			<version>1.5.22</version>
+		</dependency>
+		<!--END influxdb-client-java-->
+
+
+		<dependency>
+			<groupId>com.influxdb</groupId>
+			<artifactId>flux-dsl</artifactId>
+			<version>1.6.0</version>
+		</dependency>
+
+		<!--END InfluxDB 2.0-->
+
+		<!--START InfluxDB 1.0-->
+		<dependency>
+			<groupId>org.influxdb</groupId>
+			<artifactId>influxdb-java</artifactId>
+			<version>2.2</version>
+		</dependency>
+		<dependency>
+			<groupId>com.squareup.okhttp</groupId>
+			<artifactId>okhttp</artifactId>
+			<version>2.4.0</version>
+		</dependency>
+		<dependency>
+			<groupId>com.squareup.retrofit</groupId>
+			<artifactId>retrofit</artifactId>
+			<version>1.9.0</version>
+		</dependency>
+		<!-- END InfluxDB 1.0 -->
+	</dependencies>
+
 </project>
