+++ bundles/org.openhab.persistence.influxdb/src/main/feature/feature.xml
@@ -1,11 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<features name="org.openhab.persistence.influxdb-${project.version}" xmlns="http://karaf.apache.org/xmlns/features/v1.4.0">
-    <repository>mvn:org.openhab.core.features.karaf/org.openhab.core.features.karaf.openhab-core/${ohc.version}/xml/features</repository>
+<features name="org.openhab.persistence.influxdb-${project.version}"
+		  xmlns="http://karaf.apache.org/xmlns/features/v1.4.0">
+	<repository>
+		mvn:org.openhab.core.features.karaf/org.openhab.core.features.karaf.openhab-core/${ohc.version}/xml/features
+	</repository>
 
-    <feature name="openhab-persistence-influxdb" description="InfluxSB Persistence" version="${project.version}">
-        <feature>openhab-runtime-base</feature>
-        <bundle start-level="80">mvn:org.openhab.addons.bundles/org.openhab.persistence.influxdb/${project.version}</bundle>
-        <configfile finalname="${openhab.conf}/services/influxdb.cfg" override="false">mvn:${project.groupId}/openhab-addons-external3/${project.version}/cfg/influxdb</configfile>
-    </feature>
+	<feature name="openhab-persistence-influxdb" description="InfluxDB Persistence" version="${project.version}">
+		<feature>openhab-runtime-base</feature>
+		<bundle start-level="80">mvn:org.openhab.addons.bundles/org.openhab.persistence.influxdb/${project.version}
+		</bundle>
+		<configfile finalname="${openhab.conf}/services/influxdb.cfg" override="false">
+			mvn:${project.groupId}/openhab-addons-external3/${project.version}/cfg/influxdb
+		</configfile>
+	</feature>
 
 </features>
