+++ bundles/org.openhab.persistence.influxdb/src/main/java/org/openhab/persistence/influxdb/internal/influx1/InfluxDB1RepositoryImpl.java
@@ -1,0 +1,249 @@
+/**
+ * Copyright (c) 2010-2020 Contributors to the openHAB project
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.openhab.persistence.influxdb.internal.influx1;
+
+import static org.openhab.persistence.influxdb.internal.InfluxDBConstants.COLUMN_TIME_NAME;
+import static org.openhab.persistence.influxdb.internal.InfluxDBConstants.COLUMN_VALUE_NAME;
+
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+
+import org.eclipse.jdt.annotation.NonNullByDefault;
+import org.eclipse.jdt.annotation.Nullable;
+import org.influxdb.InfluxDB;
+import org.influxdb.InfluxDBFactory;
+import org.influxdb.dto.Point;
+import org.influxdb.dto.Pong;
+import org.influxdb.dto.Query;
+import org.influxdb.dto.QueryResult;
+import org.openhab.persistence.influxdb.InfluxRow;
+import org.openhab.persistence.influxdb.internal.InfluxDBConfiguration;
+import org.openhab.persistence.influxdb.internal.InfluxDBRepository;
+import org.openhab.persistence.influxdb.internal.InfluxPoint;
+import org.openhab.persistence.influxdb.internal.UnnexpectedConditionException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import retrofit.RetrofitError;
+
+/**
+ * Implementation of {@link InfluxDBRepository} for InfluxDB 1.0
+ *
+ * @author Joan Pujol Espinar - Initial contribution
+ */
+@NonNullByDefault
+public class InfluxDB1RepositoryImpl implements InfluxDBRepository {
+    private final Logger logger = LoggerFactory.getLogger(InfluxDB1RepositoryImpl.class);
+    private InfluxDBConfiguration configuration;
+    @Nullable
+    private InfluxDB client;
+
+    public InfluxDB1RepositoryImpl(InfluxDBConfiguration configuration) {
+        this.configuration = configuration;
+    }
+
+    /**
+     * Returns if the client has been successfully connected to server
+     *
+     * @return True if it's connected, otherwise false
+     */
+    @Override
+    public boolean isConnected() {
+        return client != null;
+    }
+
+    /**
+     * Connect to InfluxDB server
+     *
+     * @return True if successful, otherwise false
+     */
+    @Override
+    public boolean connect() {
+        final InfluxDB createdClient = InfluxDBFactory.connect(configuration.getUrl(), configuration.getUser(),
+                configuration.getPassword());
+        createdClient.enableBatch(200, 100, TimeUnit.MILLISECONDS);
+        this.client = createdClient;
+        return true;
+    }
+
+    /**
+     * Disconnect from InfluxDB server
+     */
+    @Override
+    public void disconnect() {
+        this.client = null;
+    }
+
+    /**
+     * Check if connection is currently ready
+     *
+     * @return True if its ready, otherwise false
+     */
+    @Override
+    public boolean checkConnectionStatus() {
+        boolean dbStatus = false;
+        final InfluxDB currentClient = client;
+        if (currentClient != null) {
+            try {
+                Pong pong = currentClient.ping();
+                String version = pong.getVersion();
+                // may be check for version >= 0.9
+                if (version != null && !version.contains("unknown")) {
+                    dbStatus = true;
+                    logger.debug("database status is OK, version is {}", version);
+                } else {
+                    logger.error("database ping error, version is: \"{}\" response time was \"{}\"", version,
+                            pong.getResponseTime());
+                    dbStatus = false;
+                }
+            } catch (RuntimeException e) {
+                dbStatus = false;
+                logger.error("database connection failed", e);
+                handleDatabaseException(e);
+            }
+        } else {
+            logger.error("checkConnection: database is not connected");
+        }
+        return dbStatus;
+    }
+
+    private void handleDatabaseException(Exception e) {
+        if (e instanceof RetrofitError) {
+            // e.g. raised if influxdb is not running
+            logger.error("database connection error {}", e.getMessage());
+        } else if (e instanceof RuntimeException) {
+            // e.g. raised by authentication errors
+            logger.error("database error: {}", e.getMessage());
+        }
+    }
+
+    /**
+     * Write point to database
+     *
+     * @param point
+     */
+    @Override
+    public void write(InfluxPoint point) {
+        final InfluxDB currentClient = this.client;
+        if (currentClient != null) {
+            Point clientPoint = convertPointToClientFormat(point);
+            currentClient.write(configuration.getDatabaseName(), configuration.getRetentionPolicy(), clientPoint);
+        } else {
+            logger.error("Write point {} ignored due to client isn't connected", point);
+        }
+    }
+
+    private Point convertPointToClientFormat(InfluxPoint point) {
+        Point.Builder clientPoint = Point.measurement(point.getMeasurementName()).time(point.getTime().toEpochMilli(),
+                TimeUnit.MILLISECONDS);
+        setPointValue(point.getValue(), clientPoint);
+        point.getTags().entrySet().forEach(e -> clientPoint.tag(e.getKey(), e.getValue()));
+        point.getFields().entrySet().forEach(e -> clientPoint.addField(e.getKey(), e.getValue()));
+        return clientPoint.build();
+    }
+
+    private void setPointValue(@Nullable Object value, Point.Builder point) {
+        if (value instanceof String)
+            point.addField(COLUMN_VALUE_NAME, (String) value);
+        else if (value instanceof Number)
+            point.addField(COLUMN_VALUE_NAME, (Number) value);
+        else if (value instanceof Boolean)
+            point.addField(COLUMN_VALUE_NAME, (Boolean) value);
+        else if (value == null)
+            point.addField(COLUMN_VALUE_NAME, (String) null);
+        else
+            throw new UnnexpectedConditionException("Not expected value type");
+    }
+
+    /**
+     * Executes Flux query
+     *
+     * @param query Query
+     * @return Query results
+     */
+    @Override
+    public List<InfluxRow> query(String query) {
+        final InfluxDB currentClient = client;
+        if (currentClient != null) {
+            Query parsedQuery = new Query(query, configuration.getDatabaseName());
+            List<QueryResult.Result> results = currentClient.query(parsedQuery, TimeUnit.MILLISECONDS).getResults();
+            return convertClientResutToRepository(results);
+        } else {
+            logger.error("Returning empty list because queryAPI isn't present");
+            return Collections.emptyList();
+        }
+    }
+
+    private List<InfluxRow> convertClientResutToRepository(List<QueryResult.Result> results) {
+        List<InfluxRow> rows = new ArrayList<>();
+        for (QueryResult.Result result : results) {
+            List<QueryResult.Series> seriess = result.getSeries();
+            if (result.getError() != null) {
+                logger.error("{}", result.getError());
+                continue;
+            }
+            if (seriess == null) {
+                logger.debug("query returned no series");
+            } else {
+                for (QueryResult.Series series : seriess) {
+                    logger.trace("series {}", series.toString());
+                    String itemName = series.getName();
+                    List<List<Object>> valuess = series.getValues();
+                    if (valuess == null) {
+                        logger.debug("query returned no values");
+                    } else {
+                        List<String> columns = series.getColumns();
+                        logger.trace("columns {}", columns);
+                        if (columns != null) {
+                            Integer timestampColumn = null;
+                            Integer valueColumn = null;
+                            for (int i = 0; i < columns.size(); i++) {
+                                String columnName = columns.get(i);
+                                if (columnName.equals(COLUMN_TIME_NAME)) {
+                                    timestampColumn = i;
+                                } else if (columnName.equals(COLUMN_VALUE_NAME)) {
+                                    valueColumn = i;
+                                }
+                            }
+                            if (valueColumn == null || timestampColumn == null) {
+                                throw new IllegalStateException("missing column");
+                            }
+                            for (int i = 0; i < valuess.size(); i++) {
+                                Double rawTime = (Double) valuess.get(i).get(timestampColumn);
+                                Instant time = Instant.ofEpochMilli(rawTime.longValue());
+                                Object value = valuess.get(i).get(valueColumn);
+                                logger.trace("adding historic item {}: time {} value {}", itemName, time, value);
+                                rows.add(new InfluxRow(time, itemName, value));
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        return rows;
+    }
+
+    /**
+     * Return all stored item names with it's count of stored points
+     *
+     * @return Map with <ItemName,ItemCount> entries
+     */
+    @Override
+    public Map<String, Integer> getStoredItemsCount() {
+        return Collections.emptyMap();
+    }
+}
