+++ bundles/org.openhab.persistence.influxdb/src/main/resources/ESH-INF/config/config.xml
