--- bundles/org.openhab.persistence.influxdb2/src/main/feature/feature.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<features name="org.openhab.persistence.influxdb2-${project.version}"
-		  xmlns="http://karaf.apache.org/xmlns/features/v1.4.0">
-	<repository>
-		mvn:org.openhab.core.features.karaf/org.openhab.core.features.karaf.openhab-core/${project.version}/xml/features
-	</repository>
-
-	<feature name="openhab-persistence-influxdb2" description="InfluxDB2 Binding" version="${project.version}">
-		<feature>openhab-runtime-base</feature>
-		<bundle start-level="80">mvn:org.openhab.addons.bundles/org.openhab.persistence.influxdb2/${project.version}
-		</bundle>
-	</feature>
-</features>
