--- bundles/org.openhab.persistence.influxdb2/src/main/java/org/openhab/persistence/influxdb2/RepositoryFactory.java
@@ -1,55 +1,0 @@
-/**
- * Copyright (c) 2010-2020 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.persistence.influxdb2;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.openhab.persistence.influxdb2.internal.FilterCriteriaQueryCreator;
-import org.openhab.persistence.influxdb2.internal.InfluxDB2Repository;
-import org.openhab.persistence.influxdb2.internal.InfluxDBConfiguration;
-import org.openhab.persistence.influxdb2.internal.UnnexpectedConditionException;
-import org.openhab.persistence.influxdb2.internal.influx1.Influx1FilterCriteriaQueryCreatorImpl;
-import org.openhab.persistence.influxdb2.internal.influx1.InfluxDB1RepositoryImpl;
-import org.openhab.persistence.influxdb2.internal.influx2.Influx2FilterCriteriaQueryCreatorImpl;
-import org.openhab.persistence.influxdb2.internal.influx2.InfluxDB2RepositoryImpl;
-
-/**
- * Factory that returns {@link InfluxDB2Repository} and {@link FilterCriteriaQueryCreator} implementations
- * depending on InfluxDB version
- *
- * @author Joan Pujol Espinar - Initial contribution
- */
-@NonNullByDefault
-public class RepositoryFactory {
-
-    public static InfluxDB2Repository createRepository(InfluxDBConfiguration influxDBConfiguration) {
-        switch (influxDBConfiguration.getVersion()) {
-            case V1:
-                return new InfluxDB1RepositoryImpl(influxDBConfiguration);
-            case V2:
-                return new InfluxDB2RepositoryImpl(influxDBConfiguration);
-            default:
-                throw new UnnexpectedConditionException("Not expected version " + influxDBConfiguration.getVersion());
-        }
-    }
-
-    public static FilterCriteriaQueryCreator createQueryCreator(InfluxDBConfiguration influxDBConfiguration) {
-        switch (influxDBConfiguration.getVersion()) {
-            case V1:
-                return new Influx1FilterCriteriaQueryCreatorImpl(influxDBConfiguration);
-            case V2:
-                return new Influx2FilterCriteriaQueryCreatorImpl();
-            default:
-                throw new UnnexpectedConditionException("Not expected version " + influxDBConfiguration.getVersion());
-        }
-    }
-}
