--- bundles/org.openhab.persistence.influxdb2/src/main/java/org/openhab/persistence/influxdb2/internal/FilterCriteriaQueryCreator.java
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) 2010-2020 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.persistence.influxdb2.internal;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.openhab.core.persistence.FilterCriteria;
-
-/**
- * Creates InfluxDB query sentence given a OpenHab persistence {@link FilterCriteria}
- *
- * @author Joan Pujol Espinar - Initial contribution
- */
-@NonNullByDefault
-public interface FilterCriteriaQueryCreator {
-    String createQuery(FilterCriteria criteria, String bucket);
-}
